requirements.txt
# Document processing libraries
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23

# OCR for scanned PDFs
pytesseract==0.3.10
pdf2image==1.16.3

# Configuration and environment
PyYAML==6.0.1
python-dotenv==1.0.0

# Anthropic API
anthropic==0.39.0

# Search and ML
scikit-learn==1.4.0

# Web interface
streamlit==1.40.0
